requirements.txt
alembic==1.7.6
click==8.0.3
dnspython==2.2.0
email-validator==1.1.3
Flask==2.0.2
Flask-Cors==3.0.10
Flask-Login==0.5.0
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.14.1
psycopg2-binary==2.9.3
python-dotenv==0.19.2
six==1.16.0
SQLAlchemy==1.4.31
Werkzeug==2.0.3
WTForms==3.0.1
